+++ R/table.R
@@ -68,6 +68,10 @@ sqliteListTablesQuery <- function(conn, schema = NULL, name = NULL) {
   if (is.null(schema)) {
     info_sql <- "(SELECT * FROM sqlite_master UNION ALL SELECT * FROM sqlite_temp_master)"
   } else {
+    schemas <- dbGetQuery(conn, "SELECT name FROM pragma_database_list")$name
+    if (!schema %in% schemas) {
+      return(SQL("SELECT '' AS name WHERE 0"))
+    }
     info_sql <- paste0("(SELECT * FROM ", dbQuoteIdentifier(conn, schema), ".sqlite_master)")
   }