+++ NEWS.md
@@ -1,10 +1,33 @@
 <!-- NEWS.md is maintained by https://fledge.cynkra.com, contributors should not edit this file -->
 
-# RSQLite 2.4.6.9000 (2026-02-06)
+# RSQLite 3.51.99.9001 (2026-05-09)
+
+## Bug fixes
+
+- `dbListObjects()` returns empty instead of throwing an error if database not known.
+
+## Chore
+
+- Deps.
+
+## Documentation
+
+- Align with SQLite version.
 
 ## fledge
 
-- CRAN release v2.4.6 (#685).
+- CRAN release v2.4.8 (#709).
+
+
+# RSQLite 3.51.99.9000 (2026-05-09)
+
+## Features
+
+- Upgrade bundled SQLite to 3.52.0 (#696).
+
+- Implement `dbListObjects()` for attached SQLite databases with schema prefix support (#689, #690).
+
+- Enable the percentile extension.
 
 ## Chore